Added Authentication, Completed Project

+++ taskflow/app.py
@@ -1,8 +1,9 @@
-from flask import Flask, render_template, redirect, request, flash, url_for, session, logging
+from flask import Flask, render_template, redirect, request, flash, url_for, session, logging, abort
 from flask_session import Session
 from model.db import db
 from taskflow.model.user import Register, User, LoginForm, TodoForm, Todo
 from flask_login import LoginManager
+
 
 app = Flask(__name__)
 # configure the SQLite database, relative to the app instance folder
@@ -22,25 +23,31 @@ Session(app)
 
 @app.route("/")
 def base():
+    if session.get("username"):
+        return redirect("/dashboard")
     return render_template('base.html')
 
 @app.route("/register", methods=['GET', 'POST'])
 def register():
+    if session.get("username"):
+        return redirect("/dashboard")
     form = Register()
     if form.validate_on_submit():
         user = User(
             username=form.username.data,
-            password=form.password.data
         )
+        user.set_password(form.password.data)
         #Adds user to the DB
         db.session.add(user)
         db.session.commit()
-
-        return redirect(url_for('login'))
+        session["username"] = request.form.get("username")
+        return redirect(url_for('dashboard'))
     return render_template('register.html', form=form)
 
 @app.route('/login', methods=['GET', 'POST'])
 def login():
+    if session.get("username"):
+        return redirect("/dashboard")
     form = LoginForm()
     if form.validate_on_submit():
         session["username"] = request.form.get("username")
@@ -77,15 +84,21 @@ def dashboard():
 @app.route("/dashboard/update/<int:id>", methods=["POST"])
 def update(id):
     todo = db.session.execute(db.select(Todo).filter_by(id=id)).scalar_one()
+    user = db.session.execute(db.select(User).filter_by(username=session.get("username"))).scalar_one()
+    if todo.user_id != user.id:
+        abort(403)
     todo.completed = not todo.completed
+
     db.session.commit()
     return redirect(url_for('dashboard'))
 
 @app.route("/dashboard/delete/<int:id>", methods=["POST"])
 def delete(id):
     task = db.session.execute(db.select(Todo).filter_by(id=id)).scalar_one()
+    user = db.session.execute(db.select(User).filter_by(username=session.get("username"))).scalar_one()
+    if task.user_id != user.id:
+        abort(403)
+
     db.session.delete(task)
     db.session.commit()
     return redirect(url_for('dashboard'))
-
-#check if the task belongs to the logged in user for the update/delete routes, can delete
